remove_weekly_edit_if_show_bpa_false

+++ app/src/main/java/com/wtb/dashTracker/ui/dialog_confirm/add_modify_purpose/ConfirmationDialogAddOrModifyPurposeViewModel.kt
@@ -22,7 +22,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
 
 @ExperimentalCoroutinesApi
-class ConfirmationDialogAddOrModifyPurposeViewModel: ListItemViewModel<ExpensePurpose>() {
-    override fun getItemFlowById(id: Int): Flow<ExpensePurpose?> = repository.getExpensePurposeFlowById(id)
-
+class ConfirmationDialogAddOrModifyPurposeViewModel : ListItemViewModel<ExpensePurpose>() {
+    override fun getItemFlowById(id: Long): Flow<ExpensePurpose?> =
+        repository.getExpensePurposeFlowById(id)
 }
